requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
numpy==1.26.0
faiss-cpu==1.7.4
transformers==4.35.0
torch==2.1.0
dotenv-python==0.0.1
python-dotenv==1.0.0
pymupdf==1.23.5
pytesseract==0.3.10
pillow==10.1.0
python-docx==1.0.1
requests==2.31.0
streamlit==1.28.0 